--- plugins/transport-nio/src/main/java/org/elasticsearch/transport/nio/TcpReadHandler.java
@@ -1,48 +1,0 @@
-/*
- * Licensed to Elasticsearch under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.elasticsearch.transport.nio;
-
-import org.elasticsearch.common.bytes.BytesReference;
-import org.elasticsearch.nio.NioSocketChannel;
-
-import java.io.IOException;
-
-public class TcpReadHandler {
-
-    private final String profile;
-    private final NioTransport transport;
-
-    TcpReadHandler(String profile, NioTransport transport) {
-        this.profile = profile;
-        this.transport = transport;
-    }
-
-    public void handleMessage(BytesReference reference, TcpNioSocketChannel channel, int messageBytesLength) {
-        try {
-            transport.messageReceived(reference, channel);
-        } catch (IOException e) {
-            handleException(channel, e);
-        }
-    }
-
-    public void handleException(NioSocketChannel channel, Exception e) {
-        transport.exceptionCaught(channel, e);
-    }
-}
