+++ plugins/transport-nio/src/test/java/org/elasticsearch/transport/nio/NioTransportIT.java
@@ -128,5 +128,4 @@ public class NioTransportIT extends NioIntegTestCase {
         }
 
     }
-
 }
